+++ ni-measurement-ui-creator/src/ni_measurement_ui_creator/constants/_ui_elements.py
@@ -155,7 +155,6 @@ class UpdateUI:
         "ChannelPathSelector",
         "HmiGraphPlot",
     ]
-    NUMERIC_AND_STRING_ARRAY = ["ChannelArrayNumericText", "ChannelArrayStringControl"]
 
 
 class ElementAttrib: